pipeline/requirements.txt
# Supply Chain Brain — top-level deps. See requirements.pinned.txt for
# version-bounded recommendations validated together on Python 3.14.
# Optional libs (XGBoost / lifelines / sksurv / econml / neo4j / gremlinpython)
# are detected at runtime — the application degrades gracefully without them.

# Core
pyodbc>=5,<6
pandas>=2.1,<4
numpy>=1.26,<3
sqlalchemy>=2,<3
azure-identity>=1.15,<2
msal>=1.26,<2
pyyaml>=6,<7
requests>=2.31,<3
tabulate>=0.9,<1
python-dotenv>=1,<2
openpyxl>=3.1,<4
playwright>=1.40,<2

# UI
streamlit>=1.32,<2
plotly>=5.18,<7
python-pptx>=0.6.23,<2

# Analytics core
scikit-learn>=1.3,<2
scipy>=1.10,<2
networkx>=3.1,<4

# Optional ML imputation / boosting (graceful import in brain.imputation)
# pip install xgboost lightgbm catboost missingpy
# Optional survival analysis (Phase 3.5): pip install lifelines scikit-survival
# Optional causal inference (Phase 3.1):  pip install econml
# Optional graph backends (Phase 2):      pip install neo4j gremlinpython
# Optional auth (Phase 4):                pip install streamlit-authenticator msal

